refactor(console): move api/portal/plans/edit to /api/component/plan
