+++ gravitee-apim-console-webui/src/management/api/component/plan/1-general-step/plan-edit-general-step.component.scss
