+++ gravitee-apim-console-webui/src/management/api/component/plan/3-restriction-step/plan-edit-restriction-step.component.html
