+++ gravitee-apim-console-webui/src/management/api/component/plan/api-portal-plan-edit.component.html
